style: Add .editorconfig
One might argue, that IDE related files should be ignored,
but the editorconfig is here to ensure that all code contributions
meet a certain requirement. The .editorconfig is supported by
many different IDEs and there exist a log of plugins for most IDEs.

For more info visit http://editorconfig.org