feat: Add booklist
The appearance of this first draft follows the design of the
starter project from Udacity.
In order to make the book covers appear as in the start project
we had to fetch the image and get the image's widht and height
so that we could set the corresponding styles of the cover div.
This commit happened to be a larger one, because the foundation
of the project's structure had to be laid.