fix: Memory leak on Book component
Due to a callback on image load setState() was invoked
even after the component was unmounted.
The callback is necessary to retrieve the book cover image
dimensions. But once a book is not going to get rendered anymore
the retrieval and setting of state should be stopped.
There might be some anit-pattern in the current implementation,
but at least it should be robust now.